+++ src/main/java/at/hannibal2/skyhanni/config/features/combat/damageindicator/DamageIndicatorConfig.kt
@@ -131,11 +131,6 @@ class DamageIndicatorConfig {
     var timeToKillSlayer: Boolean = true
 
     @Expose
-    @ConfigOption(name = "Time to Kill Message", desc = "Sends time to kill in Chat on Slayer Boss Death.")
-    @ConfigEditorBoolean
-    var timeToKillSlayerMessage: Boolean = true
-
-    @Expose
     @ConfigOption(name = "Show Bacte Phase", desc = "Show the current phase of Bacte in the Rift.")
     @ConfigEditorBoolean
     var showBactePhase: Boolean = true
